CIV-5944 : Refactor ccd to cui translation and fix inline variables warnings
