+++ src/main/common/models/claim.ts
@@ -39,8 +39,6 @@ import {Interest} from 'form/models/interest/interest';
 import {RejectAllOfClaimType} from 'common/form/models/rejectAllOfClaimType';
 import {ClaimDetails} from 'common/form/models/claim/details/claimDetails';
 import {ClaimantResponse} from './claimantResponse';
-import {CCDClaim} from 'models/civilClaimResponse';
-import {toCUIParty} from 'services/translation/convertToCUI/convertToCUIParty';
 import {SelfEmployedAs} from 'models/selfEmployedAs';
 import {TaxPayments} from 'models/taxPayments';
 import {RegularIncome} from 'form/models/statementOfMeans/expensesAndIncome/regularIncome';
@@ -49,11 +47,8 @@ import {CourtOrders} from 'form/models/statementOfMeans/courtOrders/courtOrders'
 import {PriorityDebts} from 'form/models/statementOfMeans/priorityDebts';
 import {Debts} from 'form/models/statementOfMeans/debts/debts';
 import {ClaimBilingualLanguagePreference} from './claimBilingualLanguagePreference';
-import {toCUIEvidence} from 'services/translation/convertToCUI/convertToCUIEvidence';
-import {toCUIClaimDetails} from 'services/translation/convertToCUI/convertToCUIClaimDetails';
 import {analyseClaimType, claimType} from 'common/form/models/claimType';
 import {PaymentIntention} from 'form/models/admission/paymentIntention';
-import {toCUIMediation} from 'services/translation/convertToCUI/convertToCUIMediation';
 
 export class Claim {
   legacyCaseReference: string;
@@ -89,16 +84,6 @@ export class Claim {
   respondent1ResponseDate?: Date;
   claimBilingualLanguagePreference: ClaimBilingualLanguagePreference;
   id: string;
-
-  public static fromCCDCaseData(ccdClaim: CCDClaim): Claim {
-    const claim: Claim = Object.assign(new Claim(), ccdClaim);
-    claim.claimDetails = toCUIClaimDetails(ccdClaim);
-    claim.evidence = toCUIEvidence(ccdClaim?.speclistYourEvidenceList);
-    claim.applicant1 = toCUIParty(ccdClaim?.applicant1);
-    claim.respondent1 = toCUIParty(ccdClaim?.respondent1);
-    claim.mediation = toCUIMediation(ccdClaim?.respondent1LiPResponse?.respondent1MediationLiPResponse);
-    return claim;
-  }
 
   get responseStatus(): ClaimResponseStatus {
     if (this.isFullAdmission() && this.isFAPaymentOptionPayImmediately()) {
