+++ src/main/services/translation/convertToCUI/cuiTranslation.ts
@@ -1,0 +1,16 @@
+import {Claim} from 'models/claim';
+import {CCDClaim} from 'models/civilClaimResponse';
+import {toCUIClaimDetails} from 'services/translation/convertToCUI/convertToCUIClaimDetails';
+import {toCUIEvidence} from 'services/translation/convertToCUI/convertToCUIEvidence';
+import {toCUIParty} from 'services/translation/convertToCUI/convertToCUIParty';
+import {toCUIMediation} from 'services/translation/convertToCUI/convertToCUIMediation';
+
+export const translateCCDCaseDatatoCUIModel = (ccdClaim: CCDClaim): Claim => {
+  const claim: Claim = Object.assign(new Claim(), ccdClaim);
+  claim.claimDetails = toCUIClaimDetails(ccdClaim);
+  claim.evidence = toCUIEvidence(ccdClaim?.speclistYourEvidenceList);
+  claim.applicant1 = toCUIParty(ccdClaim?.applicant1);
+  claim.respondent1 = toCUIParty(ccdClaim?.respondent1);
+  claim.mediation = toCUIMediation(ccdClaim?.respondent1LiPResponse?.respondent1MediationLiPResponse);
+  return claim;
+};
